+++ sdk/identity/azure-identity/inc/azure/identity/managed_identity_credential.hpp
@@ -168,6 +168,18 @@ namespace Azure { namespace Identity {
      * it was configured.
      */
     ManagedIdentityId IdentityId;
+
+    /**
+     * @brief If Azure Instance Metadata Service (IMDS) gets selected as managed identity source,
+     * specifies whether the first request should be a short probe request (`true`), instead of a
+     * normal request with retries and exponential backoff (`false`). Default is `true`.
+     *
+     * @note When `true`, there's a potential that the credential would not detect IMDS being
+     * available on a machine, if the response was not received fast enough. When `false` and IMDS
+     * is not available, credential creation may take tens of seconds until multiple attempts to get
+     * a response from IMDS would fail.
+     */
+    bool IsProbeEnabled = true;
   };
 
   /**
@@ -195,9 +207,39 @@ namespace Azure { namespace Identity {
      * @param options Options for token retrieval.
      */
     explicit ManagedIdentityCredential(
+        std::string const& clientId,
+        Core::Credentials::TokenCredentialOptions const& options);
+
+    /**
+     * @brief Constructs a Managed Identity Credential.
+     *
+     * @param clientId Client ID.
+     * @param options Options for token retrieval.
+     */
+    explicit ManagedIdentityCredential(
+        char const* clientId,
+        Core::Credentials::TokenCredentialOptions const& options);
+
+    /**
+     * @brief Constructs a Managed Identity Credential.
+     *
+     * @param clientId Client ID.
+     * @param isProbeEnabled if Azure Instance Metadata Service (IMDS) gets selected as managed
+     * identity source, specifies whether the first request should be a short probe request
+     * (`true`), instead of a normal request with retries and exponential backoff (`false`). Default
+     * is `true`.
+     * @param options Options for token retrieval.
+     *
+     * @note When @a isProbeEnabled is `true`, there's a potential that the credential would not
+     * detect IMDS being available on a machine, if the response was not received fast enough. When
+     * `false` and IMDS is not available, credential creation may take tens of seconds until
+     * multiple attempts to get a response from IMDS would fail.
+     */
+    explicit ManagedIdentityCredential(
         std::string const& clientId = std::string(),
-        Azure::Core::Credentials::TokenCredentialOptions const& options
-        = Azure::Core::Credentials::TokenCredentialOptions());
+        bool isProbeEnabled = true,
+        Core::Credentials::TokenCredentialOptions const& options
+        = Core::Credentials::TokenCredentialOptions());
 
     /**
      * @brief Constructs a Managed Identity Credential.
@@ -212,8 +254,24 @@ namespace Azure { namespace Identity {
      *
      * @param options Options for token retrieval.
      */
+    explicit ManagedIdentityCredential(Core::Credentials::TokenCredentialOptions const& options);
+
+    /**
+     * @brief Constructs a Managed Identity Credential.
+     *
+     * @param options Options for token retrieval.
+     * @param isProbeEnabled if Azure Instance Metadata Service (IMDS) gets selected as managed
+     * identity source, specifies whether the first request should be a short probe request
+     * (`true`), instead of a normal request with retries and exponential backoff (`false`).
+     *
+     * @note When @a isProbeEnabled is `true`, there's a potential that the credential would not
+     * detect IMDS being available on a machine, if the response was not received fast enough. When
+     * `false` and IMDS is not available, credential creation may take tens of seconds until
+     * multiple attempts to get a response from IMDS would fail.
+     */
     explicit ManagedIdentityCredential(
-        Azure::Core::Credentials::TokenCredentialOptions const& options);
+        bool isProbeEnabled,
+        Core::Credentials::TokenCredentialOptions const& options);
 
     /**
      * @brief Gets an authentication token.