+++ sdk/identity/azure-identity/src/managed_identity_credential.cpp
@@ -14,7 +14,8 @@ std::unique_ptr<_detail::ManagedIdentitySource> CreateManagedIdentitySource(
     std::string const& clientId,
     std::string const& objectId,
     std::string const& resourceId,
-    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    Azure::Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
 {
   using namespace Azure::Core::Credentials;
   using namespace Azure::Identity::_detail;
@@ -23,7 +24,8 @@ std::unique_ptr<_detail::ManagedIdentitySource> CreateManagedIdentitySource(
       std::string const& clientId,
       std::string const& objectId,
       std::string const& resourceId,
-      TokenCredentialOptions const& options)
+      TokenCredentialOptions const& options,
+      bool isProbeEnabled)
       = {AppServiceV2019ManagedIdentitySource::Create,
          AppServiceV2017ManagedIdentitySource::Create,
          CloudShellManagedIdentitySource::Create,
@@ -34,7 +36,8 @@ std::unique_ptr<_detail::ManagedIdentitySource> CreateManagedIdentitySource(
   // For that reason, it is not possible to cover that execution branch in tests.
   for (auto create : managedIdentitySourceCreate)
   {
-    if (auto source = create(credentialName, clientId, objectId, resourceId, options))
+    if (auto source
+        = create(credentialName, clientId, objectId, resourceId, options, isProbeEnabled))
     {
       return source;
     }
@@ -50,10 +53,25 @@ ManagedIdentityCredential::~ManagedIdentityCredential() = default;
 ManagedIdentityCredential::ManagedIdentityCredential(
     std::string const& clientId,
     Azure::Core::Credentials::TokenCredentialOptions const& options)
+    : ManagedIdentityCredential(clientId, true, options)
+{
+}
+
+ManagedIdentityCredential::ManagedIdentityCredential(
+    char const* clientId,
+    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    : ManagedIdentityCredential(std::string(clientId), options)
+{
+}
+
+ManagedIdentityCredential::ManagedIdentityCredential(
+    std::string const& clientId,
+    bool isProbeEnabled,
+    Azure::Core::Credentials::TokenCredentialOptions const& options)
     : TokenCredential("ManagedIdentityCredential")
 {
   m_managedIdentitySource
-      = CreateManagedIdentitySource(GetCredentialName(), clientId, {}, {}, options);
+      = CreateManagedIdentitySource(GetCredentialName(), clientId, {}, {}, options, isProbeEnabled);
 }
 
 ManagedIdentityCredential::ManagedIdentityCredential(
@@ -64,20 +82,20 @@ ManagedIdentityCredential::ManagedIdentityCredential(
   switch (idType)
   {
     case ManagedIdentityIdKind::SystemAssigned:
-      m_managedIdentitySource
-          = CreateManagedIdentitySource(GetCredentialName(), {}, {}, {}, options);
+      m_managedIdentitySource = CreateManagedIdentitySource(
+          GetCredentialName(), {}, {}, {}, options, options.IsProbeEnabled);
       break;
     case ManagedIdentityIdKind::ClientId:
       m_managedIdentitySource = CreateManagedIdentitySource(
-          GetCredentialName(), options.IdentityId.GetId(), {}, {}, options);
+          GetCredentialName(), options.IdentityId.GetId(), {}, {}, options, options.IsProbeEnabled);
       break;
     case ManagedIdentityIdKind::ObjectId:
       m_managedIdentitySource = CreateManagedIdentitySource(
-          GetCredentialName(), {}, options.IdentityId.GetId(), {}, options);
+          GetCredentialName(), {}, options.IdentityId.GetId(), {}, options, options.IsProbeEnabled);
       break;
     case ManagedIdentityIdKind::ResourceId:
       m_managedIdentitySource = CreateManagedIdentitySource(
-          GetCredentialName(), {}, {}, options.IdentityId.GetId(), options);
+          GetCredentialName(), {}, {}, options.IdentityId.GetId(), options, options.IsProbeEnabled);
       break;
     default:
       throw std::invalid_argument(
@@ -88,7 +106,14 @@ ManagedIdentityCredential::ManagedIdentityCredential(
 
 ManagedIdentityCredential::ManagedIdentityCredential(
     Azure::Core::Credentials::TokenCredentialOptions const& options)
-    : ManagedIdentityCredential(std::string(), options)
+    : ManagedIdentityCredential(std::string(), true, options)
+{
+}
+
+ManagedIdentityCredential::ManagedIdentityCredential(
+    bool isProbeEnabled,
+    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    : ManagedIdentityCredential(std::string(), isProbeEnabled, options)
 {
 }