+++ sdk/identity/azure-identity/src/managed_identity_source.cpp
@@ -263,8 +263,10 @@ std::unique_ptr<ManagedIdentitySource> AppServiceV2017ManagedIdentitySource::Cre
     std::string const& clientId,
     std::string const& objectId,
     std::string const& resourceId,
-    Core::Credentials::TokenCredentialOptions const& options)
+    Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
 {
+  static_cast<void>(isProbeEnabled);
   return AppServiceManagedIdentitySource::Create<AppServiceV2017ManagedIdentitySource>(
       credName, clientId, objectId, resourceId, options, "MSI_ENDPOINT", "MSI_SECRET", "2017");
 }
@@ -274,8 +276,10 @@ std::unique_ptr<ManagedIdentitySource> AppServiceV2019ManagedIdentitySource::Cre
     std::string const& clientId,
     std::string const& objectId,
     std::string const& resourceId,
-    Core::Credentials::TokenCredentialOptions const& options)
+    Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
 {
+  static_cast<void>(isProbeEnabled);
   return AppServiceManagedIdentitySource::Create<AppServiceV2019ManagedIdentitySource>(
       credName,
       clientId,
@@ -292,8 +296,10 @@ std::unique_ptr<ManagedIdentitySource> CloudShellManagedIdentitySource::Create(
     std::string const& clientId,
     std::string const& objectId,
     std::string const& resourceId,
-    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    Azure::Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
 {
+  static_cast<void>(isProbeEnabled);
   using Azure::Core::Credentials::AuthenticationException;
 
   constexpr auto EndpointVarName = "MSI_ENDPOINT";
@@ -370,8 +376,10 @@ std::unique_ptr<ManagedIdentitySource> AzureArcManagedIdentitySource::Create(
     std::string const& clientId,
     std::string const& objectId,
     std::string const& resourceId,
-    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    Azure::Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
 {
+  static_cast<void>(isProbeEnabled);
   using Azure::Core::Credentials::AuthenticationException;
 
   constexpr auto EndpointVarName = "IDENTITY_ENDPOINT";
@@ -499,7 +507,8 @@ std::unique_ptr<ManagedIdentitySource> ImdsManagedIdentitySource::Create(
     std::string const& clientId,
     std::string const& objectId,
     std::string const& resourceId,
-    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    Azure::Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
 {
   const std::string ImdsName = "Azure Instance Metadata Service";
 
@@ -529,8 +538,8 @@ std::unique_ptr<ManagedIdentitySource> ImdsManagedIdentitySource::Create(
   }
   imdsUrl.SetPath("metadata/identity/oauth2/token");
 
-  return std::unique_ptr<ManagedIdentitySource>(
-      new ImdsManagedIdentitySource(clientId, objectId, resourceId, imdsUrl, options));
+  return std::unique_ptr<ManagedIdentitySource>(new ImdsManagedIdentitySource(
+      clientId, objectId, resourceId, imdsUrl, options, isProbeEnabled));
 }
 
 ImdsManagedIdentitySource::ImdsManagedIdentitySource(
@@ -538,7 +547,8 @@ ImdsManagedIdentitySource::ImdsManagedIdentitySource(
     std::string const& objectId,
     std::string const& resourceId,
     Azure::Core::Url const& imdsUrl,
-    Azure::Core::Credentials::TokenCredentialOptions const& options)
+    Azure::Core::Credentials::TokenCredentialOptions const& options,
+    bool isProbeEnabled)
     : ManagedIdentitySource(clientId, std::string(), options),
       m_request(Azure::Core::Http::HttpMethod::Get, imdsUrl)
 {
@@ -569,7 +579,7 @@ ImdsManagedIdentitySource::ImdsManagedIdentitySource(
   Core::Credentials::TokenCredentialOptions firstRequestOptions = options;
   firstRequestOptions.Retry.MaxRetries = 0;
   m_firstRequestPipeline = std::make_unique<TokenCredentialImpl>(firstRequestOptions);
-  m_firstRequestSucceeded = false;
+  m_firstRequestSucceeded = !isProbeEnabled;
 }
 
 Azure::Core::Credentials::AccessToken ImdsManagedIdentitySource::GetToken(